Change API to satisfies(SPDX expression, array of licenses)
The design decision to take the argument expressing acceptable license
policy as another SPDX expression has repeatedly confused people.
Meanwhile, the primary use case for this package is to check some
SPDX license expression for a package against a list of approved
licenses.  I believe we can better serve that use case and make this
package easier to maintain by taking a list of approved licenses instead
of a second SPDX expression.